added PropTypes and prop type checking. Moved projectData to the utils folder as it wasn't really a component.
