+++ src/components/Project.jsx
@@ -1,8 +1,8 @@
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faGithub } from '@fortawesome/free-brands-svg-icons';
+import PropTypes from 'prop-types';
 
-
-function Project({props}){
+function Project({projects}){
 // WHEN I am presented with the Portfolio section
 // THEN I see titled images of six of the developer’s applications with links to both the deployed applications and the corresponding GitHub repository
 
@@ -14,15 +14,15 @@ function Project({props}){
 
     return(
         <div className='row'>
-            {props.map((props, index) => (
+            {projects.map((project, index) => (
                 <div key={index} className='col-md-4 mb-4'>
                      <div className="card text-bg-dark">
-                        <img src="..." className="card-img" alt={props.title} />
+                        <img src="..." className="card-img" alt={project.title} />
                         <div className="card-img-overlay">
                             <span>
-                                <h5 className="card-title">{props.title}</h5>
+                                <h5 className="card-title">{project.title}</h5>
                                 <a 
-                                    href={props.gitHubLink}
+                                    href={project.gitHubLink}
                                     target="_blank"
                                     rel="noopener noreferrer"
                                     className="text-white mx-2"
@@ -31,7 +31,7 @@ function Project({props}){
                                 </a>
                             </span>
                                 <a 
-                                    href={props.deployedLink}
+                                    href={project.deployedLink}
                                     target="_blank"
                                     rel="noopener noreferrer"
                                     className="text-white mx-2"
@@ -45,23 +45,16 @@ function Project({props}){
             ))}
         </div>
     );
-}
+};
 
-/*        <div className="card text-bg-dark">
-            <img src="..." className="card-img" alt="this will be an image" />
-        <div className="card-img-overlay">
-            <span>
-                <h5 className="card-title">Card title</h5>
-                <a 
-                        href="https://github.com/Lauren245"
-                        target="_blank"
-                        rel="noopener noreferrer"
-                        className="text-white mx-2"
-                >
-                    <FontAwesomeIcon icon={faGithub} size="2x" />
-                </a>
-            </span>
-            <p className="card-text">Languages Here</p>
-        </div>
-        </div> */
+    Project.propTypes = {
+        projects: PropTypes.arrayOf(
+            PropTypes.shape({
+                title: PropTypes.string.isRequired,
+                gitHubLink: PropTypes.string.isRequired,
+                deployedLink: PropTypes.string.isRequired,
+            })
+        ).isRequired,
+    };
+
 export default Project;
