--- src/components/ProjectData.jsx
@@ -1,47 +1,0 @@
-
-/*project card requirements:
-    1. a titled image
-    2. text of languages used
-    3.  GitHub link 
-    4. deployed link*/
-const projectsArr = [
-    {
-        title: "Cars and Trucks", 
-        skills: "TypeScript, Inquirer.js", 
-        gitHubLink: "https://github.com/Lauren245/Cars-and-Trucks", 
-        deployedLink: "#"
-    },
-    {
-        title: "Employee Tracker CMS", 
-        skills: "TypeScript, Inquirer.js, PostgreSQL", 
-        gitHubLink: "https://github.com/Lauren245/Employee-Tracker-CMS",
-        deployedLink: "#" 
-    },
-    {
-        title: "Weather Dashboard",
-        skills: "TypeScript",
-        gitHubLink: "https://github.com/Lauren245/Weather-Dashboard",
-        deployedLink: "https://weather-dashboard-ktb0.onrender.com/"
-    },
-    {
-        title: "Placeholder Card 4",
-        skills: "Skills Here",
-        gitHubLink: "https://github.com/Lauren245",
-        deployedLink: "#"
-    },
-    {
-        title: "Placeholder Card 5",
-        skills: "Skills Here",
-        gitHubLink: "https://github.com/Lauren245",
-        deployedLink: "#"
-    },
-    {
-        title: "Placeholder Card 6",
-        skills: "Skills Here",
-        gitHubLink: "https://github.com/Lauren245",
-        deployedLink: "#"
-    }
-];
-
-
-export default projectsArr
