+++ src/pages/Portfolio.jsx
@@ -1,11 +1,12 @@
 import Project from "../components/Project";
-import projectsArr from '../components/ProjectData';
+import projectsArr from '../utils/projectData';
+
 
 export default function Portfolio(){
     return (
         <div>
             <h1>My Portfolio</h1>
-            <Project props={projectsArr}/>
+            <Project projects={projectsArr}/>
         </div>
     );
 }
